// monitor/enums/Methods.ts
export enum Methods {
    GET = 'GET',
    POST = 'POST',
    PUT = 'PUT',
    DELETE = 'DELETE',
    PATCH = 'PATCH',
    COPY = 'COPY',
    HEAD = 'HEAD',
    OPTIONS = 'OPTIONS',
    LINK = 'LINK',
    UNLINK = 'UNLINK',
    PURGE = 'PURGE',
    LOCK = 'LOCK',
    UNLOCK = 'UNLOCK',
    PROPFIND = 'PROPFIND',
    VIEW = 'VIEW'
}